IGNITE-28828 Split several overloaded test suites to reduce the RunAll critical path
Split 12 overloaded Java test suites into balanced halves (by historical
per-class runtime) to shorten the longest RunAll suites. Test-infrastructure
only; no production code changed. Each split is an exact partition of the
original suite's class set (no test class added or removed).

New suites: IgniteBinaryCacheQueryTestSuite5/6, IgniteSpiDiscoverySelfTestSuite2,
IgniteCacheConsistencySelfTestSuite2, IgniteCacheFailoverTestSuite4/5,
IgniteCacheTestSuite14/15/16/17, IgniteSnapshotTestSuite7/8.

New TeamCity build configs for these suites must be added to the RunAll chain
for them to run in CI.
